tests/wpt/web-platform-tests/html/syntax/parsing/meta-inhead-insertion-mode.html

<!DOCTYPE html>
<html>
<head>
<title>Encoding specified in the "charset" attribute should have precedence over "content" attribute.</title>
<meta http-equiv="Content-Type" content="text/html; charset=koi8-r" charset="iso-8859-15">
<link rel="help" href="https://html.spec.whatwg.org/multipage/parsing.html#parsing-main-inhead">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
test(function () {
  assert_equals(document.characterSet, "ISO-8859-15");
}, "Encoding specified in the 'charset' attribute should have precedence over 'content' attribute.");
</script>
</body>
</html>
