LayoutTests/imported/w3c/web-platform-tests/html/syntax/parsing/meta-inhead-insertion-mode-expected.txt


FAIL Encoding specified in the 'charset' attribute should have precedence over 'content' attribute. assert_equals: expected "ISO-8859-15" but got "KOI8-R"
